requirements.txt
Django==1.7.8
django-extensions==1.3.8
django-shell-plus==1.1.5
ecdsa==0.11
netifaces==0.10.4
paramiko==1.12.4
pycrypto==2.6
pyreadline==2.0
requests==2.0.0
stevedore==0.15
virtualenv==1.9.1
virtualenv-clone==0.2.5
virtualenvwrapper==4.2
